put mft tests in with time_series, remove abandoned test file
