--- tests/model_connectors/test_time_series_mft_ets.py
@@ -1,94 +1,0 @@
-"""
-****************************************************************************************************
-:copyright (c) 2019-2020 URBANopt, Alliance for Sustainable Energy, LLC, and other contributors.
-
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without modification, are permitted
-provided that the following conditions are met:
-
-Redistributions of source code must retain the above copyright notice, this list of conditions
-and the following disclaimer.
-
-Redistributions in binary form must reproduce the above copyright notice, this list of conditions
-and the following disclaimer in the documentation and/or other materials provided with the
-distribution.
-
-Neither the name of the copyright holder nor the names of its contributors may be used to endorse
-or promote products derived from this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR
-IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
-FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR
-CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
-IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT
-OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-****************************************************************************************************
-"""
-
-import os
-import shutil
-import unittest
-from pathlib import Path
-
-from geojson_modelica_translator.geojson_modelica_translator import (
-    GeoJsonModelicaTranslator
-)
-from geojson_modelica_translator.model_connectors.time_series_ets_coupling import (
-    TimeSeriesConnectorETS
-)
-from geojson_modelica_translator.modelica.modelica_runner import ModelicaRunner
-from geojson_modelica_translator.system_parameters.system_parameters import (
-    SystemParameters
-)
-
-
-# FIXME: mft test should be inside the main ts_ets test class. This is literally just a duplicate of that
-class TimeSeriesModelConnectorSingleBuildingETSMFTTest(unittest.TestCase):
-    def setUp(self):
-        self.data_dir = os.path.join(os.path.dirname(__file__), 'data')
-        self.output_dir = os.path.join(os.path.dirname(__file__), 'output')
-
-        project_name = "time_series_ets_mft_ex1"
-        if os.path.exists(os.path.join(self.output_dir, project_name)):
-            shutil.rmtree(os.path.join(self.output_dir, project_name))
-
-        # load in the example geojson with a single offie building
-        feature_file = os.path.join(self.data_dir, "time_series_ex1.json")
-        self.gj = GeoJsonModelicaTranslator.from_geojson(feature_file)
-        # use the GeoJson translator to scaffold out the directory
-        self.gj.scaffold_directory(self.output_dir, project_name)
-
-        # load system parameter data
-        sys_params_file = os.path.join(self.data_dir, "time_series_system_params_ex1.json")
-        sys_params = SystemParameters(sys_params_file)
-
-        # now test the spawn connector (independent of the larger geojson translator
-        self.time_series = TimeSeriesConnectorETS(sys_params)
-        for b in self.gj.buildings:
-            self.time_series.add_building(b)
-
-    def test_timeSeries_init(self):
-        self.assertIsNotNone(self.time_series)
-        # FIXME: Better way of looking up mft data than index position 1
-        self.assertEqual(
-            self.time_series.system_parameters.get_param("buildings.custom")[1]["load_model"], "time_series"
-        )
-
-    def test_mft_time_series_to_modelica_and_run(self):
-        self.time_series.to_modelica(self.gj.scaffold)
-
-        mr = ModelicaRunner()
-        file_to_run = os.path.abspath(
-            os.path.join(
-                self.gj.scaffold.loads_path.files_dir, 'B5a6b99ec37f4de7f94020091', 'MassFlowTemperaturesTimeSeries.mo'
-            )
-        )
-        run_path = Path(os.path.abspath(self.gj.scaffold.project_path)).parent
-        exitcode = mr.run_in_docker(file_to_run, run_path=run_path, project_name=self.gj.scaffold.project_name)
-        self.assertEqual(0, exitcode)
-
-        results_path = os.path.join(run_path, f"{self.gj.scaffold.project_name}_results")
-        self.assertTrue(os.path.join(results_path, 'stdout.log'))
